% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.R
\name{is_spatial}
\alias{is_spatial}
\title{Does this stand carry stem coordinates}
\usage{
is_spatial(x)
}
\arguments{
\item{x}{A `stand`.}
}
\value{
Logical scalar.
}
\description{
Does this stand carry stem coordinates
}
